Change usize to u64 for all variables that aren't indices into byte-based structures (#34)
This changes all types that index bits instead of words to u64, adjusts the necessary conversions, and suppresses truncation warnings where truncations are impossible. 
Furthermore, it also modifies the usage, documentation, and implementation of iterator types, because various iterator functions are limited to `usize`. We avoid truncation where possible and panic where it cannot be avoided (which only applies to len() and count())